+++ projetADGR/resources/views/pages/Equipe/index.blade.php
@@ -11,7 +11,7 @@
     </div>
     <div class="row">
         <div class="col-md-12">
-            <table width="100%" class="table table-striped table-bordered table-hover dataTables-example">
+            <table width="100%" class="table table-striped">
                 <thead>
                 <th>Evenement</th>
                 <th>Membres</th>
@@ -19,25 +19,51 @@
                 <th>Actions</th>
                 </thead>
                 <tbody>
-                @foreach(\App\Equipe::all() as $equipe)
-                    <tr>
-                        {{--<td></td>--}}
-                        <td>{{$equipe->evenement->libelle}}</td>
-                        <td>
-                            @foreach($equipe->benevoleEquipe as $be)
-                                <b>-</b> <a href="/benevole/show/{{$be->benevole->id}}">{{$be->benevole->nom." ".$be->benevole->prenom}}</a><br>
-                            @endforeach
-                        </td>
-                        <td><a href="/benevole/show/{{$equipe->responsable->id}}">{{$equipe->responsable->nom." ".$equipe->responsable->prenom}}</a></td>
-                        <td>
-                            <a href="/equipe/delete/{{$equipe->id}}"><span class=" btn btn-warning btn-circle btn-md glyphicon glyphicon-remove removeCollecte"></span></a>
-                            <a href="/equipe/edit/{{$equipe->id}}"><span class=" btn btn-default btn-circle btn-md glyphicon glyphicon-pencil"></span></a>
-                            <a href="/equipe/show/{{$equipe->id}}"><span class=" btn btn-default btn-circle btn-md glyphicon glyphicon-list"></span></a>
-                        </td>
-                    </tr>
-                @endforeach
+                <?php
+                    $equipes = \App\Equipe::paginate(10);
+                ?>
+                @if(Auth::user()->role->id == 1)
+                    @foreach($equipes as $equipe)
+                        <tr>
+                            {{--<td></td>--}}
+                            <td>{{$equipe->evenement->libelle}}</td>
+                            <td>
+                                @foreach($equipe->benevoleEquipe as $be)
+                                    <b>-</b> <a href="/benevole/show/{{$be->benevole->id}}">{{$be->benevole->nom." ".$be->benevole->prenom}}</a><br>
+                                @endforeach
+                            </td>
+                            <td><a href="/benevole/show/{{$equipe->responsable->id}}">{{$equipe->responsable->nom." ".$equipe->responsable->prenom}}</a></td>
+                            <td>
+                                <a href="/equipe/delete/{{$equipe->id}}"><span class=" btn btn-warning btn-circle btn-md glyphicon glyphicon-remove removeCollecte"></span></a>
+                                <a href="/equipe/edit/{{$equipe->id}}"><span class=" btn btn-default btn-circle btn-md glyphicon glyphicon-pencil"></span></a>
+                                <a href="/equipe/show/{{$equipe->id}}"><span class=" btn btn-default btn-circle btn-md glyphicon glyphicon-list"></span></a>
+                            </td>
+                        </tr>
+                    @endforeach
+                @else
+                    @foreach($equipes as $equipe)
+                        @if($equipe->responsable->zone->ville->id == Auth::user()->zone->ville->id)
+                            <tr>
+                                {{--<td></td>--}}
+                                <td>{{$equipe->evenement->libelle}}</td>
+                                <td>
+                                    @foreach($equipe->benevoleEquipe as $be)
+                                        <b>-</b> <a href="/benevole/show/{{$be->benevole->id}}">{{$be->benevole->nom." ".$be->benevole->prenom}}</a><br>
+                                    @endforeach
+                                </td>
+                                <td><a href="/benevole/show/{{$equipe->responsable->id}}">{{$equipe->responsable->nom." ".$equipe->responsable->prenom}}</a></td>
+                                <td>
+                                    <a href="/equipe/delete/{{$equipe->id}}"><span class=" btn btn-warning btn-circle btn-md glyphicon glyphicon-remove removeCollecte"></span></a>
+                                    <a href="/equipe/edit/{{$equipe->id}}"><span class=" btn btn-default btn-circle btn-md glyphicon glyphicon-pencil"></span></a>
+                                    <a href="/equipe/show/{{$equipe->id}}"><span class=" btn btn-default btn-circle btn-md glyphicon glyphicon-list"></span></a>
+                                </td>
+                            </tr>
+                        @endif
+                    @endforeach
+                @endif
                 </tbody>
             </table>
+            {{$equipes->links()}}
         </div>
     </div>
 </div>
